+++ js/get-random-entity.js
